+++ org.emoflon.neo.neo4j.adapter/META-INF/MANIFEST.MF
@@ -23,7 +23,6 @@ Export-Package: org.emoflon.neo.cypher.common,
  org.emoflon.neo.cypher.models.templates,
  org.emoflon.neo.cypher.patterns,
  org.emoflon.neo.cypher.rules,
- org.emoflon.neo.neo4j.adapter,
  org.neo4j.driver.internal,
  org.neo4j.driver.internal.async,
  org.neo4j.driver.internal.async.inbound,
